Import GCM data to R

This function imports GCM stack files from folder to R.

Usage

import_gcms(
  path = "input_data/WorldClim_data_gcms",
  extension = ".tif",
  recursive = TRUE,
  gcm_names = NULL,
  var_names = NULL
)

Arguments

path

A string with the path to GCM files.

extension

Extension of stack files. Standard is ".tif", which is the extension from WorldClim 2.1.

recursive

Logical. Should the function import stacks recursively (i.e. search files from folders within folders)? Standard is TRUE.

gcm_names

A vector with names to be addressed to each GCM.

var_names

A vector with names to be addressed to each variable.

Value

A list of stacks, with each element of the list corresponding to a GCM from given path.

See also

Author

Luíz Fernando Esser (luizesser@gmail.com) https://luizfesser.wordpress.com

Examples

if (FALSE) { # \dontrun{
s <- import_gcms(
  path = "input_data/WorldClim_data_future", extension = ".tif",
  recursive = TRUE, gcm_names = NULL
)
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
t <- transform_gcms(s, var_names, study_area)
} # }