Summarize General Circulation Model (GCM) data

This function summarizes GCM data by calculating several statistics for each variable

Usage

summary_gcms(s, var_names = c("bio_1", "bio_12"), study_area = NULL)

Arguments

s

A list of stacks of General Circulation Models.

var_names

Character. A vector with names of the bioclimatic variables to compare OR 'all'.

study_area

Extent object, or any object from which an Extent object can be extracted. A object that defines the study area for cropping and masking the rasters.

Value

a data frame with the summary statistics for each variable

See also

Author

Luíz Fernando Esser (luizesser@gmail.com) https://luizfesser.wordpress.com

Examples

if (FALSE) { # \dontrun{
s <- list(stack("gcm1.tif"), stack("gcm2.tif"), stack("gcm3.tif"))
study_area <- extent(c(-57, -22, -48, -33))
var_names <- c("bio_1", "bio_12")
s <- transform_gcms(s, var_names, study_area)
summary_gcms(s)
} # }